fix noisy neighbor tests
